Adding to the Person data model.
Broke up Person class into several partial classes. Trying to prevent a monolithic class from emerging.
Added a second fake person, which will be used in testing Parents and Children later on.
Added logic to Add or Remove Parents and Children.
Added DateOfBirth and DateOfDeath to Person, need this to make a unique person.
Working on Equality logic for Person.