Added tests for PersonFamily and fixed a bug with Email equality
Turns out that Email equality would fail whenever the Email list is empty. This is an issue since the ContactInformation class will instantiate the Email list on construction. MailAddress will return false if the email that you are comparing from is null. This is because it will then leverage ToString().Equals which also will return false if the object is null.